pyproject.toml
[tool.poetry]
name = "leases-licensing"
version = "1.0.0"
description = "Leases Licensing"
authors = ["DBCA OIM <asi@dbca.wa.gov.au>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2"
django-confy = "^1.0" # old
dj-database-url = "^0.5" # old
django_countries = "^7.3"
django_cron = "^0.6"
django-reversion = "^5.0"
django-reversion-compare = "^0.15"
geojson = "^2.5"
python_dateutil = "^2.8"
pytz = "^2020.1"
reportlab = "^3.5"
requests = "^2.3"
setuptools = "^45.2"
six = "^1.0"
# git+https://github.com/xzzy/webtemplate.git#egg=webtemplate_dbca
webtemplate_dbca = {git = "https://github.com/dbca-wa/webtemplate2.git"}
wagov_utils = {git = "https://github.com/dbca-wa/wagov_utils.git"}
# git+https://github.com/dbca-wa/dbca-utils@1.1.1#egg=dbca-utils
dbca-utils = {git = "https://github.com/dbca-wa/dbca-utils.git", rev = "1.1.1"}
psycopg2-binary = "^2.0"
django-extensions = "^3.1"
# git+https://github.com/xzzy/ledger_api_client.git#egg=ledger_api_client
ledger_api_client = {git = "https://github.com/dbca-wa/ledger_api_client.git"}
whitenoise = "^6.0"
djangorestframework = "^3.14"
djangorestframework-gis = "1.0"
openpyxl = "^3.0"
django-ckeditor = "^6.0"
django-widget-tweaks = "^1.4"
djangorestframework-datatables = "^0.7"
django-utils-six = "^2.0" # old
django-phonenumber-field = "^6.1"
phonenumbers = "^8.12"
babel = "^2.10"
django-dirtyfields = "^1.8"
django_media_serv = {git = "https://github.com/dbca-wa/django-media-serv.git", rev="main"}
pandas = "1.5.2"
geopandas = "^0.12"
python-docx = "^0.8"
gunicorn = "^19.10"
drf-standardized-errors = "^0.12"
pre-commit = "^3.4.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.4"
ipython = "^8.15"
ipdb = "^0.13"
pylint = "^2.13"
flake8 = "^6.1"
autopep8 = "^2.0"
black = "^23.0"

[build-system]
requires = ["poetry_core>=1.1.13"]
build-backend = "poetry.core.masonry.api"
